Refactor accessibility fallback search pattern
Instead of using `while keepSearching`, we now
have a `findElement(at)` method that is recursive.
It follows the same principal as in it will continue
to search for a window by decrementing the y coordinate
until it will eventually hit the toolbar or end up
returning `nil`.

It also has a global `isSearching` boolean that will
help to stop unnecessary lookup. If the `state` is
`.ended`, any ongoing searches should stop.
This improves reliability, performance and most importantly
it makes the user experience more seemless.